src/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
psycopg2-binary==2.9.10
PyJWT==2.10.0
sqlparse==0.5.2
